add bood deletion update

+++ src/app/pages/user/list-books/list-books.component.ts
@@ -2,6 +2,7 @@ import { Component, Input } from '@angular/core';
 import { ResponseBook } from '../../../core/interfaces/book.interface';
 import { BookService } from '../../../core/service/book/book.service';
 import { ChatBoxCommunicateService } from '../../../core/service/chat/chat-box-communicate.service';
+import { SnackbarService } from '../../../core/service/shared/snackbar/snackbar.service';
 
 @Component({
   selector: 'app-list-books',
@@ -13,13 +14,20 @@ export class ListBooksComponent {
 
   constructor(
     private bookService : BookService,
-    private comService : ChatBoxCommunicateService
+    private comService : ChatBoxCommunicateService,
+    private snBarService : SnackbarService
     ){}
 
   deleteBook(id : string){
+    if(!confirm("Deleting the book")) return;
     this.bookService.deleteBook(id).subscribe({
       next : res =>{
         console.log(res)
+        this.books = this.books.filter(each => each.id !== id);
+        this.snBarService.openSnackBar("book deleted success fully")
+      },
+      error : err => {
+        this.snBarService.openSnackBar(err.message.message || "failed to delete")
       }
     })
   }
